Scripts/requirements_automation.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
google-cloud-bigquery==3.13.0
google-auth==2.23.4
requests==2.31.0 